PCNT, POPM, POSS, RBUF, RVAD, RVRB tags

+++ src/main/java/com/aurawin/mp3/payload/BufferSize.java
@@ -3,5 +3,5 @@ package com.aurawin.mp3.payload;
 public class BufferSize {
     public long Size;
     public byte Flags;
-    public int Offset;
+    public long Offset;
 }
